feat: update deps

+++ webpack/webpack-start-server-plugin/src/start-server.plugin.ts
@@ -1,6 +1,6 @@
 import { ChildProcess }      from 'node:child_process'
 import { fork }              from 'node:child_process'
-import { resolve }           from 'node:path'
+import { join }              from 'node:path'
 import { Writable }          from 'node:stream'
 
 import webpack               from 'webpack'
@@ -20,6 +20,8 @@ export class StartServerPlugin {
 
   worker: ChildProcess | null = null
 
+  initialized: boolean = false
+
   logger: StartServerLogger
 
   constructor(options: Partial<StartServerPluginOptions> = {}) {
@@ -28,86 +30,39 @@ export class StartServerPlugin {
       entryName: 'index',
       ...options,
     }
-
-    this.enableRestarting()
-  }
-
-  private enableRestarting(): void {
-    process.stdin.setEncoding('utf8')
-    process.stdin.on('data', (data: string) => {
-      if (data.trim() === 'rs') {
-        this.restartServer()
-      }
-    })
-  }
-
-  getEntryFile(compilation) {
-    const { entryName } = this.options
-    const { entrypoints } = compilation
-
-    const entry = entrypoints.get ? entrypoints.get(entryName) : entrypoints[entryName!]
-
-    if (!entry) {
-      throw new Error(
-        `Requested entry "${entryName}" does not exist, try one of: ${(entrypoints.keys
-          ? entrypoints.keys()
-          : Object.keys(entrypoints)
-        ).join(' ')}`
-      )
-    }
-
-    const entryScript = webpack.EntryPlugin
-      ? // eslint-disable-next-line no-underscore-dangle
-        entry._runtimeChunk.files.values().next().value
-      : entry.chunks[0].files[0]
-
-    if (!entryScript) {
-      throw new Error(`Entry chunk not outputted: ${entry}`)
-    }
-
-    const { path } = compilation.outputOptions
-
-    return resolve(path, entryScript)
   }
 
   private afterEmit = (compilation: webpack.Compilation, callback: () => void): void => {
-    if (this.worker && this.worker.connected && this.worker?.pid) {
-      this.restartServer()
-      callback()
-      return
-    }
+    if (!this.initialized) {
+      this.initialized = true
 
-    this.startServer(compilation, callback)
+      callback()
+    } else {
+      if (this.worker?.connected && this.worker?.pid) {
+        process.kill(this.worker.pid)
+      }
+
+      this.startServer(compilation, callback)
+    }
   }
 
   apply = (compiler: webpack.Compiler): void => {
     compiler.hooks.afterEmit.tapAsync({ name: 'StartServerPlugin' }, this.afterEmit)
   }
 
-  private restartServer(): void {
-    this.logger.info('Restarting service...')
+  private startServer = (compilation: webpack.Compilation, callback: () => void): void => {
+    this.logger.info('Start server...')
 
-    if (this.worker?.pid) {
-      process.kill(this.worker.pid)
-    }
-
-    if (this.entryFile) {
-      this.runWorker(this.entryFile, (worker) => {
-        this.worker = worker
-      })
-    }
-  }
-
-  private startServer = (compilation: webpack.Compilation, cb: () => void): void => {
-    this.entryFile = this.getEntryFile(compilation)
+    this.entryFile = join(compilation.compiler.options.output.path!, 'index.js')
 
     this.runWorker(this.entryFile, (worker) => {
       this.worker = worker
-      cb()
+
+      callback()
     })
   }
 
-  private runWorker(entryFile: string, cb: (arg0: ChildProcess) => void): void {
+  private runWorker(entryFile: string, callback: (arg0: ChildProcess) => void): void {
     const worker = fork(entryFile, [], {
       silent: true,
     })
@@ -120,6 +75,6 @@ export class StartServerPlugin {
       worker.stderr?.pipe(this.options.stderr, { end: false })
     }
 
-    setTimeout(() => cb(worker), 0)
+    setTimeout(() => callback(worker), 0)
   }
 }
